Refactor code to store diagnostic information with optimizations
