+++ src/ac_hemi_test.h
@@ -6,22 +6,22 @@
 #ifndef Racmacs__ac_hemi_test__h
 #define Racmacs__ac_hemi_test__h
 
-struct HemiDiagnosis
+class HemiDiagnosis
 {
+
+public:
   std::string diagnosis;
   arma::vec coords;
+
 };
 
-struct HemiData
+class HemiData
 {
+
+public:
   arma::uword index;
   std::vector<HemiDiagnosis> diagnoses;
-};
 
-struct HemiTestOutput
-{
-  std::vector<HemiData> antigens;
-  std::vector<HemiData> sera;
 };
 
 #endif
