+++ src/acmap_diagnostics.h
@@ -1,0 +1,15 @@
+
+#include <RcppArmadillo.h>
+#include "ac_hemi_test.h"
+
+#ifndef Racmacs__acmap_diagnostics__h
+#define Racmacs__acmap_diagnostics__h
+
+class AcDiagnostics {
+
+  public:
+    std::vector<HemiDiagnosis> hemi;
+
+};
+
+#endif
